Rename PayPalDataCollector to FraudProtection
