+++ Demo/build.gradle
@@ -61,7 +61,7 @@ android {
 
 dependencies {
     debugImplementation project(':Card')
-    debugImplementation project(':PayPalDataCollector')
+    debugImplementation project(':FraudProtection')
     debugImplementation project(':PayPalNativeCheckout')
     debugImplementation project(':PayPalWebCheckout')
     debugImplementation project(':PayPalUI')
