+++ FraudProtection/proguard-rules.pro
