+++ FraudProtection/src/main/java/com/paypal/android/fraudprotection/UUIDHelper.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.paypaldatacollector
+package com.paypal.android.fraudprotection
 
 import android.content.Context
 import java.util.UUID
